add data augmentation argument & minor fixes

+++ utils/training.py
@@ -9,7 +9,7 @@ def train(epoch_index, loader, model, criterion, optimizer, scaler, device, wand
     loss_accumulator = 0
     with tqdm(loader, desc=f"Training epoch {epoch_index}") as progress_container:
         for batch_index, (images, masks, _, _) in enumerate(progress_container):
-            images = images.to(device=device)
+            images = images.float().to(device=device)
             masks = masks.squeeze(dim=1).long().to(device=device)
 
             # Forward
